refactor: Split large files into focused modules under 500 lines
- notifications.py: 516 lines → 13 lines (main) + 2 focused modules
  - toast_notifier.py: 52 lines (cross-platform notifications)
  - weather_notifier.py: 472 lines (alert processing & persistence)

- gui/ui/display_managers.py: 515 lines → 13 lines (main) + 4 focused modules
  - state_manager.py: 60 lines (loading/ready states)
  - forecast_display.py: 198 lines (forecast formatting)
  - conditions_display.py: 235 lines (current conditions)
  - error_handler.py: 84 lines (error display)

- services/update_service.py: 510 lines → 13 lines (main) + 4 focused modules
  - update_info.py: 38 lines (data structures)
  - version_checker.py: 237 lines (version comparison & checking)
  - download_manager.py: 117 lines (download & installation)

✅ All modules now under 500 lines (target achieved)
✅ Maintained backward compatibility via __init__.py imports
✅ Improved separation of concerns and maintainability
✅ Preserved all existing functionality and interfaces

Total reduction: 1541 lines → 39 lines in main files + focused modules
